Button to return to main menu after quiz
There is now a button in the finished quiz menu that lets users return to the main menu after taking the quiz. The addition of this button required major refactoring as I realised that QuizRunnerModel was directly modifying the reference dictionaries in IQuestionService. IQuestionService now returns a deep copy instead of a shallow copy, which has solved this problem.